+++ packages/api/pocketbase/pb_migrations/1769198401_updated_users.js
@@ -1,0 +1,67 @@
+/// <reference path="../pb_data/types.d.ts" />
+migrate((app) => {
+  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+
+  // add field
+  collection.fields.addAt(10, new Field({
+    "autogeneratePattern": "",
+    "hidden": false,
+    "id": "text4049598648",
+    "max": 0,
+    "min": 0,
+    "name": "morningStart",
+    "pattern": "",
+    "presentable": false,
+    "primaryKey": false,
+    "required": false,
+    "system": false,
+    "type": "text"
+  }))
+
+  // add field
+  collection.fields.addAt(11, new Field({
+    "autogeneratePattern": "",
+    "hidden": false,
+    "id": "text3506237662",
+    "max": 0,
+    "min": 0,
+    "name": "afternoonStart",
+    "pattern": "",
+    "presentable": false,
+    "primaryKey": false,
+    "required": false,
+    "system": false,
+    "type": "text"
+  }))
+
+  // add field
+  collection.fields.addAt(12, new Field({
+    "autogeneratePattern": "",
+    "hidden": false,
+    "id": "text3885690521",
+    "max": 0,
+    "min": 0,
+    "name": "eveningStart",
+    "pattern": "",
+    "presentable": false,
+    "primaryKey": false,
+    "required": false,
+    "system": false,
+    "type": "text"
+  }))
+
+  return app.save(collection)
+}, (app) => {
+  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+
+  // remove field
+  collection.fields.removeById("text4049598648")
+
+  // remove field
+  collection.fields.removeById("text3506237662")
+
+  // remove field
+  collection.fields.removeById("text3885690521")
+
+  return app.save(collection)
+})